Elton/Adding in some colliders for the water (need to find a way to make it not hardcoded as of now). Added in simple conjure which just puts the sprite in a set location. Current goals: -Scrolling camera -Being able to conjure based on where your mouse is. -More generalized pet that is not necessarily a baked in prefab.
