+++ Assets/ElvGames/Fantasy Dreamland/Sprites/Characters/Materials.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 990e5decfaafe1d4eb59c3d0f152c3fb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
