+++ Assets/ElvGames/Fantasy Dreamland/Sprites/Characters/Materials/Character_041.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 51e7a5153b39d0e41a3595e5b7e227f8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
