+++ Assets/ElvGames/Fantasy Dreamland/Sprites/Characters/Materials/Character_097.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9b38af9135d94ac4eb443f3016cae9b4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
